Ensure rejection based on origin AS0 is done after accepting based on other ASNs
This is somewhat twisted: the existence of one valid RPKI ROA should not
preclude the existence of another valid RPKI ROA. However, if there no
ROAs except a ROA with AS 0 as origin, the intent is to reject any route
announcements. If there multiple ROAs, including one that has AS 0 as
origin, the effect of the AS 0 ROA is nullified by using a deterministic
reverse sort.